backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.4",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "greenlet>=3.3.2",
    "nicegui>=3.8.0",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
